implemented all basic functions

+++ airogami/common/com/airogami/common/constants/ChainConstants.java
@@ -4,6 +4,6 @@ public interface ChainConstants {
 
 	public final byte StatusUnmatched = 0;//not found next member
 	public final byte StatusMatched = 1;// found next member
-	public final byte MaxMatchCount = 20;
-	public final byte MaxPassCount = 20;
+	public final short MaxMatchCount = 20;
+	public final short MaxPassCount = 20;
 }
